more diff between operations and commands
