+++ webdriver/src/commands.ts
@@ -7,26 +7,6 @@ import type { EventBusInterface } from "./eventbus.js";
 
 let headers = new Headers([["Content-Type", "application/json"]]);
 
-export async function webdriverCommands(
-	req: IncomingMessage,
-	res: ServerResponse,
-	params: WebdriverParams,
-) {
-	let { url } = params;
-	let urlStr = url.toString();
-	if (urlStr === "/cmd/find_element") {
-		// findElement request
-	}
-	if (urlStr === "/cmd/element_click") {
-	}
-	if (urlStr === "/cmd/element_send_keys") {
-	}
-	if (urlStr === "/cmd/send_keys") {
-	}
-	if (urlStr === "/cmd/take_element_screenshot") {
-	}
-}
-
 export async function newSession(params: WebdriverParams, signal: AbortSignal) {
 	let { url, capabilities } = params;
 
@@ -77,6 +57,93 @@ export async function deleteSession(
 		});
 	}
 }
+
+export async function go(
+	params: WebdriverParams,
+	signal: AbortSignal,
+	sessionId: string,
+	hostAndPort: URL,
+	targetPath: string,
+) {
+	let { url } = params;
+
+	let pingUrl = new URL(targetPath, hostAndPort);
+	let getCookie = await fetch(new URL(`/session/${sessionId}/url`, url), {
+		method: "POST",
+		headers,
+		body: JSON.stringify({ url: pingUrl }),
+		signal,
+	});
+
+	if (200 !== getCookie.status) {
+		let cause = await getCookie.json();
+		throw new Error("go-to-cookie request failed", { cause });
+	}
+}
+
+
+export async function addCookie(
+	params: WebdriverParams,
+	signal: AbortSignal,
+	sessionId: string,
+) {
+	let { url, jackrabbitId } = params;
+
+	let cookieReq = await fetch(new URL(`/session/${sessionId}/cookie`, url), {
+		method: "POST",
+		headers,
+		body: JSON.stringify({
+			cookie: {
+				name: "jackrabbit",
+				value: jackrabbitId,
+				path: "/",
+				httpOnly: true,
+			},
+		}),
+		signal,
+	});
+
+	if (200 !== cookieReq.status) {
+		let cause = await cookieReq.json();
+		throw new Error("set-cookie request failed", { cause });
+	}
+}
+
+// need event bus to send errors to error log
+export async function findElement(
+	params: WebdriverParams, // driver defined state
+	signal: AbortSignal | undefined, // driver defined state
+	sessionId: string, // derived state associated with driver
+	cssSelector: string,
+): Promise<string | undefined> {
+	let { url } = params;
+
+	let res = await fetch(
+		new URL(new URL(`/session/${sessionId}/element`, url)),
+		{
+			method: "GET",
+			headers,
+			body: JSON.stringify({ using: "css selector", value: cssSelector }),
+			signal,
+		},
+	);
+
+	if (200 === res.status) {
+		let json = await res.json();
+		if ("object" !== typeof json?.value)
+			throw new Error("getElements return value is not an object");
+
+		for (let [key, value] of json?.value.entries()) {
+			if (
+				"string" === typeof key &&
+				"string" === typeof value &&
+				key.startsWith("element-")
+			)
+				return value;
+		}
+	}
+}
+
 
 function getJsonFromRequestBody(req: IncomingMessage): Promise<any> {
 	return new Promise(function (resolve, reject) {
